ai_support_system/requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Telegram bot
aiogram==3.3.0

# Embeddings & ML
sentence-transformers>=2.2.0
faiss-cpu>=1.8.0
numpy>=1.24.0,<2.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Clustering
scikit-learn==1.4.0

# Templates
jinja2==3.1.3

# Utils
python-dotenv==1.0.1
openpyxl>=3.1.0
pydantic-settings==2.1.0
httpx==0.26.0
cachetools==5.3.2
